minor refactor x2 -e bench: 1530940

+++ src/search/search.cpp
@@ -130,7 +130,7 @@ namespace episteme::search {
 
             Line candidate = {};
             int32_t score = 0;
-            int16_t search_depth = depth - 1;
+            int search_depth = depth - 1;
 
             if (num_legal >= 4 && depth >= 3) {
                 int16_t reduction = 1;
